VR movement works, and is now much more intuitive.
In the previous implementation, one would move in the direction the controller was pointing. Now you move in the direction your head is facing, or rather that transformed by the direction of the joystick. e.g. If you look forward and move right with the joystick, you move right. I had to make InputManager.ApplyHorizontalForce and InputManager.ApplyVerticalForce public in order to make the player move like it does in the base game.