VR menu input works, lasers work more intuitively
* VR menus can now be clicked when using headsets other than the Vive
* World objects that have their hint active with the laser can now be
selected accurately. The raycast for clicking was previously different
than the one for activating the hint, leading to frustration.
* Laser is now activated when you do a pointing gesture (hold grip and
let go of trigger)
* Laser and interaction raycasts are now cast from the pointed index
finger, rather than the entire hand's forward direction.
* PlayerActivate now activates from rayEmitter position (in base game:
the camera, in VR: the finger) rather than from a hardcoded height.
* player hands are visible when paused
* Cursor on worldspace canvas is a rawimage on the canvas instead of a
mesh. Its position is now accurate to where the laser is pointing.
* Changed keycode for resetting player position in DaggerfallVRPLayer to
keycode.P. Need to set this as a SteamVR input binding instead, and
maybe add it to a custom VR menu.