Implemented VR equipment system. Should work, but spawns cubes.
Currently untested in VR. To avoid having to clutter source control with big files, I'm having it load resources from the untracked folder. If weapons are not present in a VR/Equipment Resources folder, it will load whatever is in the VR/EquipmentBackup resources folder. Right now that's a cube with weapon name.